requirements.txt
Flask
flask-cors
youtube-transcript-api
google-genai
google-generativeai
python-dotenv